+++ models/phone-model.mjs
@@ -1,7 +1,9 @@
 import mongoose from 'mongoose';
 import createdAtPlugin from './helpers/created-at';
+import addCSVContentPlugin from './helpers/add-csv-content';
 
 const { Schema } = mongoose;
+const MODEL_NAME = 'PhoneModel';
 
 const PhoneModelSchema = new Schema({
   _id: String,
@@ -9,6 +11,7 @@ const PhoneModelSchema = new Schema({
 });
 
 PhoneModelSchema.plugin(createdAtPlugin);
+PhoneModelSchema.plugin(addCSVContentPlugin);
 
 PhoneModelSchema.index({
   _id: 'text',
@@ -17,4 +20,4 @@ PhoneModelSchema.index({
 
 PhoneModelSchema.statics.getDashedName = () => 'phone-model';
 
-export default mongoose.model('PhoneModel', PhoneModelSchema, 'phone-models');
+export default mongoose.model(MODEL_NAME, PhoneModelSchema, 'phone-models');