+++ routes/users.mjs
@@ -12,6 +12,7 @@ import {
   CAN_EDIT_USER_WITHOUT_UPPER_RIGHTS,
 } from '../models/rights';
 import config from '../services/config';
+import { csvToJson } from '../middlewares/csv-to-json';
 import contentNegociation from '../middlewares/content-negociation';
 import maskOutput from '../middlewares/mask-output';
 
@@ -223,6 +224,11 @@ const router = generateCRUD(User, {
         { body },
       );
     },
+  },
+  batch: {
+    right: [CAN_CREATE_USER],
+    refs: ['email'],
+    middlewares: [csvToJson],
   },
 });