+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Current
 ### Features
 * Add recurring time slots
+* Add csv middleware to list routes when needed
+* Add batch crud route to handle csv imports
 * Export metrics for prometheus
 ### Fix
 * bump validator to 12.1.0 & deleted node-input-validator
@@ -14,14 +16,13 @@
 * Add password constraint
 * Add password expiration
 * Add rights for admin to update his own campuses
-* Add csv middleware to list routes when needed
 * Add date coherence control
 * Add mergeMasks helper
 * Add defaultReservationScope to campus model
 * Add possibility for ride owner to cancel ride by updating status
+### Fix
 * Remove logs from mongodb, using Loki stack right now
 * Export dashboard and monitoring on deploy
-### Fix
 * Rides and ratings routes tests
 * Link in registration mail
 * Fix fixtures script